refactor: migrate page AboutMe
