+++ eslint.config.mjs
@@ -33,7 +33,7 @@ export default tseslint.config(
       ...reactHooks.configs.recommended.rules,
       'react-refresh/only-export-components': [
         'warn',
-        { allowConstantExport: true },
+        { allowExportNames: ['metadata'], allowConstantExport: true },
       ],
       'react-compiler/react-compiler': 'error',
       ...react.configs.recommended.rules,
