+++ postcss.config.mjs
@@ -2,7 +2,7 @@ const config = {
   plugins: {
     '@tailwindcss/postcss': {},
     autoprefixer: {},
-  }
+  },
 };
 
 export default config;
