+++ src/app/about-me/index.ts
@@ -1,0 +1,4 @@
+export { default as DescriptionAboutMe } from './description-about-me';
+export { default as ItemContact } from './item-contact';
+export { default as ItemTechnicalSkill } from './item-technical-skills';
+export { codeString, contacts, skills, softSkill } from './types/types';
